Extract VMR patch handling into a separate file (#2019)
