+++ src/Microsoft.DotNet.Darc/src/DarcLib/VirtualMonoRepo/VmrInitializer.cs
@@ -13,6 +13,11 @@ using Microsoft.Extensions.Logging;
 #nullable enable
 namespace Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 
+/// <summary>
+/// This class is able to initialize an individual repository within the VMR for the first time.
+/// It pulls in the new sources adhering to cloaking rules, accommodating for patched files, resolving submodules.
+/// It can also initialize all other repositories recursively based on the dependencies stored in Version.Details.xml.
+/// </summary>
 public class VmrInitializer : VmrManagerBase, IVmrInitializer
 {
     // Message shown when initializing an individual repo for the first time
@@ -24,18 +29,21 @@ public class VmrInitializer : VmrManagerBase, IVmrInitializer
         """;
     
     private readonly IVmrDependencyTracker _dependencyTracker;
+    private readonly IVmrPatchHandler _patchHandler;
     private readonly ILogger<VmrUpdater> _logger;
 
     public VmrInitializer(
         IVmrDependencyTracker dependencyTracker,
+        IVmrPatchHandler patchHandler,
         IProcessManager processManager,
         IRemoteFactory remoteFactory,
         IVersionDetailsParser versionDetailsParser,
         ILogger<VmrUpdater> logger,
         IVmrManagerConfiguration configuration)
-        : base(dependencyTracker, processManager, remoteFactory, versionDetailsParser, logger, configuration.TmpPath)
+        : base(dependencyTracker, patchHandler, processManager, remoteFactory, versionDetailsParser, logger, configuration.TmpPath)
     {
         _dependencyTracker = dependencyTracker;
+        _patchHandler = patchHandler;
         _logger = logger;
     }
 
@@ -60,17 +68,17 @@ public class VmrInitializer : VmrManagerBase, IVmrInitializer
         var commit = GetCommit(clone, (targetRevision is null || targetRevision == HEAD) ? null : targetRevision);
 
         string patchPath = GetPatchFilePath(mapping);
-        await CreatePatch(mapping, clonePath, Constants.EmptyGitObject, commit.Id.Sha, patchPath, cancellationToken);
+        await _patchHandler.CreatePatch(mapping, clonePath, Constants.EmptyGitObject, commit.Id.Sha, patchPath, cancellationToken);
         cancellationToken.ThrowIfCancellationRequested();
 
-        await ApplyPatch(mapping, patchPath, cancellationToken);
+        await _patchHandler.ApplyPatch(mapping, patchPath, cancellationToken);
         cancellationToken.ThrowIfCancellationRequested();
 
         _dependencyTracker.UpdateDependencyVersion(mapping, new(commit.Id.Sha, targetVersion));
         Commands.Stage(new Repository(_dependencyTracker.VmrPath), VmrDependencyTracker.GitInfoSourcesDir);
         cancellationToken.ThrowIfCancellationRequested();
 
-        await ApplyVmrPatches(mapping, cancellationToken);
+        await _patchHandler.ApplyVmrPatches(mapping, cancellationToken);
         cancellationToken.ThrowIfCancellationRequested();
 
         await UpdateGitmodules(cancellationToken);
