+++ src/Microsoft.DotNet.Darc/src/DarcLib/VirtualMonoRepo/VmrManagerBase.cs
@@ -21,10 +21,9 @@ namespace Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 public abstract class VmrManagerBase : IVmrManager
 {
     protected const string HEAD = "HEAD";
-    private const string KeepAttribute = "vmr-preserve";
-    private const string IgnoreAttribute = "vmr-ignore";
 
     private readonly IVmrDependencyTracker _dependencyInfo;
+    private readonly IVmrPatchHandler _patchHandler;
     private readonly IProcessManager _processManager;
     private readonly IRemoteFactory _remoteFactory;
     private readonly IVersionDetailsParser _versionDetailsParser;
@@ -36,6 +35,7 @@ public abstract class VmrManagerBase : IVmrManager
 
     protected VmrManagerBase(
         IVmrDependencyTracker dependencyInfo,
+        IVmrPatchHandler patchHandler,
         IProcessManager processManager,
         IRemoteFactory remoteFactory,
         IVersionDetailsParser versionDetailsParser,
@@ -44,6 +44,7 @@ public abstract class VmrManagerBase : IVmrManager
     {
         _logger = logger;
         _dependencyInfo = dependencyInfo;
+        _patchHandler = patchHandler;
         _processManager = processManager;
         _remoteFactory = remoteFactory;
         _versionDetailsParser = versionDetailsParser;
@@ -76,146 +77,6 @@ public abstract class VmrManagerBase : IVmrManager
         remoteRepo.Clone(mapping.DefaultRemote, mapping.DefaultRef, clonePath, checkoutSubmodules: false, null);
 
         return clonePath;
-    }
-
-    /// <summary>
-    /// Creates a patch file (a diff) for given two commits in a repo adhering to the in/exclusion filters of the mapping.
-    /// </summary>
-    protected async Task CreatePatch(
-        SourceMapping mapping,
-        string repoPath,
-        string sha1,
-        string sha2,
-        string destPath,
-        CancellationToken cancellationToken)
-    {
-        _logger.LogInformation("Creating diff in {path}..", destPath);
-
-        var args = new List<string>
-        {
-            "diff",
-            "--patch",
-            "--binary", // Include binary contents as base64
-            "--output", // Store the diff in a .patch file
-            destPath,
-            $"{sha1}..{sha2}",
-            "--",
-        };
-
-        if (!mapping.Include.Any())
-        {
-            mapping = mapping with
-            {
-                Include = new[] { "**/*" }
-            };
-        }
-
-        args.AddRange(mapping.Include.Select(p => $":(glob,attr:!{IgnoreAttribute}){p}"));
-        args.AddRange(mapping.Exclude.Select(p => $":(exclude,glob,attr:!{KeepAttribute}){p}"));
-
-        // Other git commands are executed from whichever folder and use `-C [path to repo]` 
-        // However, here we must execute in repo's dir because attribute filters work against the working tree
-        // We also need to do call this from the repo root and not from repo/.git
-        var result = await _processManager.Execute(
-            _processManager.GitExecutable,
-            args,
-            workingDir: repoPath.EndsWith(".git") ? Path.GetDirectoryName(repoPath) : repoPath,
-            cancellationToken: cancellationToken);
-
-        result.ThrowIfFailed($"Failed to create an initial diff for {mapping.Name}");
-
-        _logger.LogDebug("{output}", result.ToString());
-
-        args = new List<string>
-        {
-            "rev-list",
-            "--count",
-            $"{sha1}..{sha2}",
-        };
-
-        var distance = (await _processManager.ExecuteGit(repoPath, args, cancellationToken)).StandardOutput.Trim();
-
-        _logger.LogInformation("Diff created at {path} - {distance} commit{s}, {size}",
-            destPath, distance, distance == "1" ? string.Empty : "s", StringUtils.GetHumanReadableFileSize(destPath));
-    }
-
-    /// <summary>
-    /// Applies a given patch file onto given mapping's subrepository.
-    /// </summary>
-    /// <param name="mapping">Mapping</param>
-    /// <param name="patchPath">Path to the patch file with the diff</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    protected async Task ApplyPatch(SourceMapping mapping, string patchPath, CancellationToken cancellationToken)
-    {
-        // We have to give git a relative path with forward slashes where to apply the patch
-        var destPath = _dependencyInfo.GetRepoSourcesPath(mapping)
-            .Replace(_dependencyInfo.VmrPath, null)
-            .Replace("\\", "/")
-            [1..];
-
-        _logger.LogInformation("Applying patch {patchPath} to {path}...", patchPath, destPath);
-
-        // This will help ignore some CR/LF issues (e.g. files with both endings)
-        (await _processManager.ExecuteGit(_dependencyInfo.VmrPath, new[] { "config", "apply.ignoreWhitespace", "change" }, cancellationToken: cancellationToken))
-            .ThrowIfFailed("Failed to set git config whitespace settings");
-
-        Directory.CreateDirectory(destPath);
-
-        IEnumerable<string> args = new[]
-        {
-            "apply",
-
-            // Apply diff to index right away, not the working tree
-            // This is faster when we don't care about the working tree
-            // Additionally works around the fact that "git apply" failes with "already exists in working directory"
-            // This happens only when case sensitive renames happened in the history
-            // More details: https://lore.kernel.org/git/YqEiPf%2FJR%2FMEc3C%2F@camp.crustytoothpaste.net/t/
-            "--cached",
-
-            // Options to help with CR/LF and similar problems
-            "--ignore-space-change",
-
-            // Where to apply the patch into
-            "--directory",
-            destPath,
-
-            patchPath,
-        };
-
-        var result = await _processManager.ExecuteGit(_dependencyInfo.VmrPath, args, cancellationToken: CancellationToken.None);
-        result.ThrowIfFailed($"Failed to apply the patch for {destPath}");
-        _logger.LogDebug("{output}", result.ToString());
-
-        // After we apply the diff to the index, working tree won't have the files so they will be missing
-        // We have to reset working tree to the index then
-        // This will end up having the working tree all staged
-        _logger.LogInformation("Resetting the working tree...");
-        args = new[] { "checkout", destPath };
-        result = await _processManager.ExecuteGit(_dependencyInfo.VmrPath, args, cancellationToken: CancellationToken.None);
-        result.ThrowIfFailed($"Failed to clean the working tree");
-        _logger.LogDebug("{output}", result.ToString());
-    }
-
-    /// <summary>
-    /// Applies VMR patches onto files of given mapping's subrepository.
-    /// </summary>
-    /// <param name="mapping">Mapping</param>
-    protected async Task ApplyVmrPatches(SourceMapping mapping, CancellationToken cancellationToken)
-    {
-        if (!mapping.VmrPatches.Any())
-        {
-            return;
-        }
-
-        _logger.LogInformation("Applying VMR patches for {mappingName}..", mapping.Name);
-
-        foreach (var patchFile in mapping.VmrPatches)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            _logger.LogDebug("Applying {patch}..", patchFile);
-            await ApplyPatch(mapping, patchFile, cancellationToken);
-        }
     }
 
     /// <summary>
